Updated Scoring
Created PositiveIntegerField for Better to hold the score.

For now, the admin is able to manipulate the scores of the players. I believe this should always be a backup option for the admin but I will implement a scoring system later.

Added HTML code for logging out of the system and Code for printing the rank, username, and score to a table. The table does not sort through yet so that needs to be implemented.

To-Do
Implement Scoring System (Talk to Dr. Bartz)
Sort the table in HTML to display proper ranking according to score
Have not yet tested with inactive users